feat: parametric generators setup
generators are now parametric so that each parameter can be overridden.

the logic is being kept as simple as possible intentionally, to validate
the approach first.

this comes at the cost of the following downsides:
- not _all_ valid generators are treated as such: tuples of generators
  are being treated as constants
- a generator is always interpreted as such, and it's not possible to
  assign a generator as a static parameter
- to override parameters of a parametric generator, such as
  interface's generator inside device's, the only solution is declaring
  the custom interface generator and then use it on the device call

these features are all possible to implement later with ad-hoc syntax or
checks, without breaking changes and keeping the happy-path simple.

Signed-off-by: Francesco Noacco <francesco.noacco@secomind.com>